+++ src/main/java/com/nhnacademy/store99/bookstore/book/repository/impl/BookRepositoryImpl.java
@@ -3,7 +3,7 @@ package com.nhnacademy.store99.bookstore.book.repository.impl;
 import com.nhnacademy.store99.bookstore.book.dto.response.BookResponse;
 import com.nhnacademy.store99.bookstore.book.entity.Book;
 import com.nhnacademy.store99.bookstore.book.entity.QBook;
-import com.nhnacademy.store99.bookstore.book.repository.BookRepository;
+import com.nhnacademy.store99.bookstore.book.repository.BookRepositoryCustom;
 import com.querydsl.core.types.Projections;
 import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport;
 import org.springframework.stereotype.Component;
@@ -12,9 +12,10 @@ import org.springframework.stereotype.Component;
  * 도서 QueryDSL 레포지토리 구현체 <br>
  *
  * @author yrrho2
+ * @author seunggyu-kim
  */
 @Component
-public class BookRepositoryImpl extends QuerydslRepositorySupport implements BookRepository {
+public class BookRepositoryImpl extends QuerydslRepositorySupport implements BookRepositoryCustom {
     public BookRepositoryImpl() {
         super(Book.class);
     }
