+++ src/main/java/com/nhnacademy/store99/bookstore/book_category/service/Impl/BookCategoryServiceImpl.java
@@ -22,7 +22,6 @@ public class BookCategoryServiceImpl implements BookCategoryService {
         return bookCategoryRepository.getCategoriesByParentsId(categoryId);
     }
 
-
     @Override
     public Page<BookListElementDTO> getBooksByCategories(Long categoryId, Pageable pageable) {
         List<CategoryParentsDTO> cp = bookCategoryRepository.getCategoriesByParentsId(categoryId);
