+++ src/main/java/com/nhnacademy/store99/bookstore/order_book/repository/impl/OrderBookRepositoryImpl.java
@@ -2,9 +2,9 @@ package com.nhnacademy.store99.bookstore.order_book.repository.impl;
 
 import com.nhnacademy.store99.bookstore.book.entity.QBook;
 import com.nhnacademy.store99.bookstore.order_book.DTO.response.LatestBookResponse;
-import com.nhnacademy.store99.bookstore.order_book.OrderBook;
+import com.nhnacademy.store99.bookstore.order_book.entity.OrderBook;
 import com.nhnacademy.store99.bookstore.order_book.entity.QOrderBook;
-import com.nhnacademy.store99.bookstore.order_book.repository.OrderBookRepository;
+import com.nhnacademy.store99.bookstore.order_book.repository.OrderBookRepositoryCustom;
 import com.querydsl.core.group.GroupBy;
 import java.util.List;
 import java.util.Map;
@@ -12,7 +12,7 @@ import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class OrderBookRepositoryImpl extends QuerydslRepositorySupport implements OrderBookRepository {
+public class OrderBookRepositoryImpl extends QuerydslRepositorySupport implements OrderBookRepositoryCustom {
     public OrderBookRepositoryImpl() {
         super(OrderBook.class);
     }
